django_projects/dwyloapp_API/requirement.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Cerberus==1.3.4
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
django-otp==1.1.3
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
importlib-metadata==4.10.1
Markdown==3.3.6
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==1.7.1
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
